# apps/api/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Movie Picker API",
    "description": "API REST pour l'application Movie Picker (soirées film, propositions, votes, roue).",
    "version": "0.0.1"
  },
  "servers": [
    { "url": "http://localhost:4000", "description": "Développement" }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Santé du service",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "service": { "type": "string", "example": "movie-picker-api" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/search": {
      "get": {
        "summary": "Recherche de films (TMDB)",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Terme de recherche"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste de résultats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "title": { "type": "string" },
                      "year": { "type": "string" },
                      "posterPath": { "type": "string", "nullable": true }
                    }
                  }
                }
              }
            }
          },
          "503": { "description": "Recherche indisponible (TMDB non configurée)" }
        }
      }
    },
    "/events": {
      "post": {
        "summary": "Créer un event",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "date", "time"],
                "properties": {
                  "title": { "type": "string" },
                  "date": { "type": "string", "example": "2025-03-15" },
                  "time": { "type": "string", "example": "20:00" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event créé (hostToken dans la réponse à conserver pour les actions hôte)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "title": { "type": "string" },
                    "date": { "type": "string" },
                    "time": { "type": "string" },
                    "slug": { "type": "string" },
                    "shareUrl": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation échouée" }
        }
      }
    },
    "/events/slug/{idOrSlug}": {
      "get": {
        "summary": "Détail d'un event par slug",
        "parameters": [
          { "name": "idOrSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "host", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Token hôte pour isHost" }
        ],
        "responses": {
          "200": {
            "description": "Event avec isHost, terminé, winnerMovie",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventDetail" }
              }
            }
          },
          "404": { "description": "Soirée introuvable" }
        }
      }
    },
    "/events/{idOrSlug}": {
      "get": {
        "summary": "Détail d'un event par id ou slug",
        "parameters": [
          { "name": "idOrSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "host", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event" },
          "404": { "description": "Soirée introuvable" }
        }
      }
    },
    "/events/{idOrSlug}/join": {
      "post": {
        "summary": "Rejoindre un event",
        "parameters": [
          { "name": "idOrSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["pseudo"],
                "properties": {
                  "pseudo": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Participant créé" },
          "200": { "description": "Déjà inscrit avec ce pseudo" },
          "400": { "description": "Soirée terminée ou validation" },
          "404": { "description": "Soirée introuvable" }
        }
      }
    },
    "/events/{idOrSlug}/wheel": {
      "post": {
        "summary": "Lancer la roue (réservé à l'hôte)",
        "parameters": [
          { "name": "idOrSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "host", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Film gagnant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "winner": { "type": "object" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Aucun film / roue déjà lancée / soirée clôturée" },
          "403": { "description": "Réservé à l'hôte" },
          "404": { "description": "Soirée introuvable" }
        }
      }
    },
    "/events/{idOrSlug}/close": {
      "post": {
        "summary": "Clôturer l'event (réservé à l'hôte)",
        "parameters": [
          { "name": "idOrSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "host", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Soirée clôturée" },
          "403": { "description": "Réservé à l'hôte" },
          "404": { "description": "Soirée introuvable" }
        }
      }
    },
    "/events/{idOrSlug}/movies": {
      "get": {
        "summary": "Liste des films de l'event",
        "parameters": [
          { "name": "idOrSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Liste des films avec proposerPseudo, score, up, down",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/MovieWithScore" }
                }
              }
            }
          },
          "404": { "description": "Soirée introuvable" }
        }
      },
      "post": {
        "summary": "Ajouter un film à l'event",
        "parameters": [
          { "name": "idOrSlug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tmdbId", "title", "year", "participantId"],
                "properties": {
                  "tmdbId": { "type": "integer" },
                  "title": { "type": "string" },
                  "year": { "type": "string" },
                  "posterPath": { "type": "string", "nullable": true },
                  "participantId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Film ajouté" },
          "400": { "description": "Soirée terminée / participant invalide" },
          "409": { "description": "Déjà proposé (même TMDB ou titre)" },
          "404": { "description": "Soirée introuvable" }
        }
      }
    },
    "/events/{idOrSlug}/movies/{movieId}": {
      "delete": {
        "summary": "Supprimer un film (par le proposant, si roue non lancée)",
        "parameters": [
          { "name": "idOrSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "movieId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["participantId"],
                "properties": {
                  "participantId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Film supprimé" },
          "400": { "description": "Roue déjà lancée" },
          "403": { "description": "Seul le proposant peut retirer" },
          "404": { "description": "Film ou soirée introuvable" }
        }
      }
    },
    "/events/{idOrSlug}/movies/{movieId}/vote": {
      "post": {
        "summary": "Voter (upvote ou downvote)",
        "parameters": [
          { "name": "idOrSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "movieId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["participantId", "value"],
                "properties": {
                  "participantId": { "type": "string" },
                  "value": { "type": "integer", "enum": [1, -1] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Vote enregistré" },
          "400": { "description": "Soirée terminée / participant invalide" },
          "404": { "description": "Film ou soirée introuvable" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EventDetail": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "date": { "type": "string" },
          "time": { "type": "string" },
          "slug": { "type": "string" },
          "isHost": { "type": "boolean" },
          "terminé": { "type": "boolean" },
          "winnerMovie": { "type": "object", "nullable": true }
        }
      },
      "MovieWithScore": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "year": { "type": "string" },
          "posterPath": { "type": "string", "nullable": true },
          "proposerPseudo": { "type": "string" },
          "score": { "type": "integer" },
          "up": { "type": "integer" },
          "down": { "type": "integer" }
        }
      }
    }
  }
}